add some adaptive and smth
