+++ social-network-ui/src/redux/slices/authSlice.js
@@ -1,0 +1,20 @@
+import { createSlice } from '@reduxjs/toolkit'
+
+const initialState = {
+	isAuthenticated: false,
+	user: null,
+}
+
+const authSlice = createSlice({
+	name: 'auth',
+	initialState,
+	reducers: {
+		login: (state, action) => {
+			state.isAuthenticated = true
+			state.user = action.payload
+		},
+	},
+})
+
+export const { login } = authSlice.actions
+export default authSlice.reducer
